+++ R/make.R
@@ -246,6 +246,12 @@
 #' Lazy loading may be more memory efficient in some use cases, but
 #' it may duplicate the loading of dependencies, costing time.
 #'
+#' @param session_info logical, whether to save the \code{sessionInfo()}
+#' to the cache. This behavior is recommended for serious \code{\link{make}()}s
+#' for the sake of reproducibility. This argument only exists to
+#' speed up tests. Apparently, \code{sessionInfo()} is a bottleneck
+#' for small \code{\link{make}()}s.
+#'
 #' @examples
 #' \dontrun{
 #' test_with_dir("Quarantine side effects.", {
@@ -308,7 +314,8 @@ make <- function(
   skip_safety_checks = FALSE,
   store_meta = TRUE,
   config = NULL,
-  lazy_load = FALSE
+  lazy_load = FALSE,
+  session_info = TRUE
 ){
   force(envir)
   if (!is.null(return_config)){
@@ -353,7 +360,8 @@ make <- function(
       skip_imports = skip_imports,
       skip_safety_checks = skip_safety_checks,
       store_meta = store_meta,
-      lazy_load = lazy_load
+      lazy_load = lazy_load,
+      session_info = session_info
     )
   }
   make_with_config(config = config)
@@ -456,9 +464,11 @@ targets_graph <- function(config){
 
 initialize_session <- function(config){
   config$cache$clear(namespace = "session")
-  config$cache$set(
-    key = "sessionInfo",
-    value = sessionInfo(),
-    namespace = "session"
-  )
+  if (config$session_info){
+    config$cache$set(
+      key = "sessionInfo",
+      value = sessionInfo(),
+      namespace = "session"
+    )
+  }
 }