+++ tests/testthat/test-full-build.R
@@ -18,8 +18,8 @@ test_with_dir("scratch build with custom filesystem cache.", {
   expect_true(is.numeric(readd(final, cache = cache)))
   expect_true(length(config$cache$list()) > 2)
   expect_false(any(c("f", "final") %in% ls()))
-  expect_true(is.list(drake_session(cache = cache)))
-  expect_true(all(drake_session(cache = cache)$target %in% config$plan$target))
+  expect_true(
+    all(read_drake_plan(cache = cache)$target %in% config$plan$target))
 
   cache <- this_cache(path = path)
   expect_equal(short_hash(cache), "murmur32")